Remove local_settings since we're no longer using that setup and merge all settings into base
