+++ mapstory/settings/base.py
@@ -27,11 +27,16 @@ import geonode
 from geonode.settings import *
 import pyproj
 
+
+def str_to_bool(v):
+    return v.lower() in ("yes", "true", "t", "1")
+
 #
 # General Django development settings
 #
 
 SITENAME = 'MapStory'
+SITEURL = "%s://%s" % (os.environ['PUBLIC_PROTOCOL'], os.environ['PUBLIC_HOST'])
 
 # Defines the directory that contains the settings file as the LOCAL_ROOT
 # It is used for relative settings elsewhere.
@@ -43,8 +48,8 @@ STATICFILES_DIRS = [
                        os.path.join(LOCAL_ROOT, "static"),
                    ] + STATICFILES_DIRS
 
-STATIC_ROOT = os.path.join(LOCAL_ROOT, "static_root")
-MEDIA_ROOT = os.path.join(LOCAL_ROOT, "uploaded")
+MEDIA_ROOT = os.environ['MEDIA_ROOT']
+STATIC_ROOT = os.environ['STATIC_ROOT']
 
 
 # Location of url mappings
@@ -111,6 +116,16 @@ INSTALLED_APPS += (
 )
 # Thanks !
 
+MAPSTORY_APPS = (
+
+    'mapstory.apps.boxes',
+    'mapstory.apps.flag', # - temporarily using this instead of the flag app for django because they need to use AUTH_USER_MODEL
+
+)
+
+INSTALLED_APPS += MAPSTORY_APPS
+
+
 # Adding Threaded Comments app
 FLUENT_COMMENTS_EXCLUDE_FIELDS = ('name', 'email', 'url', 'title')
 COMMENTS_APP = 'fluent_comments'
@@ -144,18 +159,22 @@ TEMPLATES = [
     },
 ]
 
+# Geoserver Settings
+GEOSERVER_PUBLIC_LOCATION = "%s://%s/geoserver/" % (os.environ['PUBLIC_PROTOCOL'], os.environ['PUBLIC_HOST'])
+GEOSERVER_LOCATION = "%s://%s:%d/geoserver/" % (os.environ['PRIVATE_PROTOCOL'], os.environ['GEOSERVER_HOST_INTERNAL'], int(os.environ['GEOSERVER_PORT_INTERNAL']))
+
 OGC_SERVER = {
     'default': {
         'BACKEND': 'geonode.geoserver',
-        'LOCATION': 'http://localhost:8080/geoserver/',
+        'LOCATION': GEOSERVER_LOCATION,
         'LOGIN_ENDPOINT': 'j_spring_oauth2_geonode_login',
         'LOGOUT_ENDPOINT': 'j_spring_oauth2_geonode_logout',
         # PUBLIC_LOCATION needs to be kept like this because in dev mode
         # the proxy won't work and the integration tests will fail
         # the entire block has to be overridden in the local_settings
-        'PUBLIC_LOCATION': 'http://localhost:8000/geoserver/',
+        'PUBLIC_LOCATION': GEOSERVER_PUBLIC_LOCATION,
         'USER': 'admin',
-        'PASSWORD': 'geoserver',
+        'PASSWORD': os.environ['GEOSERVER_PASSWORD'],
         'MAPFISH_PRINT_ENABLED': True,
         'PRINT_NG_ENABLED': True,
         'GEONODE_SECURITY_ENABLED': True,
@@ -173,23 +192,73 @@ OGC_SERVER = {
     }
 }
 
+# Debugging Settings
 DEBUG_STATIC = True
+DEBUG = str_to_bool(os.environ.get('DEBUG', 'False'))
+if not DEBUG:
+    ALLOWED_HOSTS = os.environ.get('ALLOWED_HOSTS', '').split('|')
+if os.environ['PUBLIC_HOST'].replace('.', '').isdigit():
+    # IP Address
+    SESSION_COOKIE_DOMAIN = os.environ['PUBLIC_HOST']
+elif '.' in os.environ['PUBLIC_HOST']:
+    # Domain name, hopefully
+    SESSION_COOKIE_DOMAIN = ".%s" % (os.environ['PUBLIC_HOST'],)
+else:
+    # hostname with no TLD?
+    SESSION_COOKIE_DOMAIN = "%s" % (os.environ['PUBLIC_HOST'],)
 
+# Registration Settings
 REGISTRATION_OPEN = True
 
 LOCAL_CONTENT = False
 
-DATABASE_PASSWORD = None
-DATABASE_HOST = 'localhost'
+DATABASE_PASSWORD = os.environ['DATABASE_PASSWORD']
+DATABASE_HOST = os.environ['DATABASE_HOST']
 DATABASE_PORT = '5432'
 
 AUTOCOMPLETE_QUICK_SEARCH = False
 
-ACCOUNT_EMAIL_CONFIRMATION_REQUIRED = True
+# Email Settings
+ACCOUNT_EMAIL_CONFIRMATION_REQUIRED = str_to_bool(os.environ['ACCOUNT_EMAIL_CONFIRMATION_REQUIRED'])
 ACCOUNT_EMAIL_CONFIRMATION_EMAIL = True
+EMAIL_HOST = os.environ.get('EMAIL_HOST', '')
+EMAIL_HOST_USER = os.environ.get('EMAIL_HOST_USER', '')
+EMAIL_HOST_PASSWORD = os.environ.get('EMAIL_HOST_PASSWORD', '')
+EMAIL_PORT = int(os.environ.get('EMAIL_PORT', '25'))
+EMAIL_USE_TLS = str_to_bool(os.environ.get('EMAIL_USE_TLS', 'false'))
+EMAIL_BACKEND = os.environ.get('EMAIL_BACKEND', 'django.core.mail.backends.console.EmailBackend')
+THEME_ACCOUNT_CONTACT_EMAIL = os.environ.get('EMAIL_HOST_USER', '')
+ACCOUNT_OPEN_SIGNUP = True
+ACCOUNT_ACTIVATION_DAYS = int(os.environ.get('ACCOUNT_ACTIVATION_DAYS', '0'))
+DEFAULT_FROM_EMAIL = os.environ.get('DEFAULT_FROM_EMAIL', '')
+ACCOUNT_EMAIL_CONFIRMATION_AUTHENTICATED_REDIRECT_URL = '/'
+ACCOUNT_LOGIN_REDIRECT_URL = '/'
+ACCOUNT_EMAIL_CONFIRMATION_ANONYMOUS_REDIRECT_URL = '/'
 
+# AWS S3 Settings
+
+# Required to run Sync Media to S3
+AWS_BUCKET_NAME = os.environ.get('AWS_STORAGE_BUCKET_NAME' ,'')
+AWS_STORAGE_BUCKET_NAME = os.environ.get('AWS_STORAGE_BUCKET_NAME','')
+AWS_ACCESS_KEY_ID = os.environ.get('AWS_ACCESS_KEY_ID','')
+AWS_SECRET_ACCESS_KEY = os.environ.get('AWS_SECRET_ACCESS_KEY','')
+AWS_S3_BUCKET_DOMAIN = '%s.s3.amazonaws.com' % (AWS_STORAGE_BUCKET_NAME,)
 USE_AWS_S3_STATIC = False
 USE_AWS_S3_MEDIA = False
+
+if USE_AWS_S3_STATIC:
+    STATICFILES_LOCATION = 'static'
+    STATICFILES_STORAGE = 'mapstory.s3_storages.StaticStorage'
+    STATIC_URL = "https://%s/%s/" % (AWS_S3_BUCKET_DOMAIN, STATICFILES_LOCATION)
+
+if USE_AWS_S3_MEDIA:
+    MEDIAFILES_LOCATION = 'media'
+    MEDIA_URL = "https://%s/%s/" % (AWS_S3_BUCKET_DOMAIN, MEDIAFILES_LOCATION)
+    DEFAULT_FILE_STORAGE = 'mapstory.s3_storages.MediaStorage'
+
+REMOTE_CONTENT_URL = STATIC_URL + 'assets'
+
+# Django OSGEO Importer Settings
 
 IMPORT_HANDLERS = (
     'mapstory.import_handlers.TruncatedNameHandler',
@@ -204,9 +273,10 @@ IMPORT_HANDLERS = (
 )
 
 OSGEO_IMPORTER_GEONODE_ENABLED = True
+OSGEO_DATASTORE = 'datastore'
+# Soft time limit for the import_object celery task of django_osgeo_importer, should be changed later after testing.
+IMPORT_TASK_SOFT_TIME_LIMIT = 1800
 
-if os.path.exists('mapstory/settings/local_settings.py'):
-    exec open('mapstory/settings/local_settings.py') in globals()
 
 # Download formats available in layer detail download modal
 DOWNLOAD_FORMATS_VECTOR = [
@@ -330,21 +400,6 @@ MAP_BASELAYERS = [
         'title': 'World Dark',
         'group': 'background'
     },
-    ## remove esri basemap until supported by storytools
-    ##    {
-    ##        "source": {"ptype": "gxp_arcrestsource",
-    ##                   "url": "https://services.arcgisonline.com/arcgis/rest/services/NGS_Topo_US_2D/MapServer/",
-    ##                   "alwaysAnonymous": True,
-    ##                   'proj': 'EPSG:4326'},
-    ##        "type": "OpenLayers.Layer",
-    ##        "args": ["Worldmap", "https://services.arcgisonline.com/arcgis/rest/services/NGS_Topo_US_2D/MapServer/",
-    ##                 {"layers": 'basic'}],
-    ##        "visibility": False,
-    ##        "fixed": True,
-    ##        "group": "background",
-    ##        "name": "NGS_Topo_US_2D",
-    ##        "title": "Esri NGS"
-    ##    },
     {
         'source': {
             'ptype': 'gxp_mapboxsource',
@@ -459,24 +514,15 @@ LOGGING = {
     },
 }
 
-if USE_AWS_S3_STATIC:
-    STATICFILES_LOCATION = 'static'
-    STATICFILES_STORAGE = 'mapstory.s3_storages.StaticStorage'
-    STATIC_URL = "https://%s/%s/" % (AWS_S3_BUCKET_DOMAIN, STATICFILES_LOCATION)
-
-if USE_AWS_S3_MEDIA:
-    MEDIAFILES_LOCATION = 'media'
-    MEDIA_URL = "https://%s/%s/" % (AWS_S3_BUCKET_DOMAIN, MEDIAFILES_LOCATION)
-    DEFAULT_FILE_STORAGE = 'mapstory.s3_storages.MediaStorage'
-
-REMOTE_CONTENT_URL = STATIC_URL + 'assets'
-
 # the layer_create view allows users to create layer by providing a workspace and a featureType
 # this settings whitelists the datastores in which layers creation are allowed
 ALLOWED_DATASTORE_LAYER_CREATE = ('*',)
 
-# Gravatar Settings
+# Avatar Settings
+AUTO_GENERATE_AVATAR_SIZES = (35, 45, 75, 100)
+AVATAR_DEFAULT_URL = "%s/static/mapstory/img/default_avatar_lg.png" % SITEURL
 AVATAR_GRAVATAR_SSL = True
+AVATAR_GRAVATAR_BACKUP = False
 
 DEFAULT_IMPORTER_CONFIG = {
     'configureTime': True,
@@ -558,3 +604,87 @@ ACTSTREAM_SETTINGS = {
     'USE_JSONFIELD': True,
     'GFK_FETCH_DEPTH': 1,
 }
+
+THEME = os.environ.get('THEME', 'default')
+USER_SNAP = True
+
+# Google Analytics Settings
+GOOGLE_ANALYTICS = os.environ.get('GOOGLE_ANALYTICS', '')
+
+# Celery Settings
+BROKER_URL = "amqp://mapstory:%s@%s/%s" % (os.environ['RABBITMQ_APPLICATION_PASSWORD'], os.environ['RABBITMQ_HOST'], os.environ['RABBITMQ_APPLICATION_VHOST'])
+CELERY_DEFAULT_QUEUE = "default"
+CELERY_DEFAULT_EXCHANGE = "default"
+CELERY_DEFAULT_EXCHANGE_TYPE = "direct"
+CELERY_DEFAULT_ROUTING_KEY = "default"
+CELERY_CREATE_MISSING_QUEUES = True
+CELERY_ALWAYS_EAGER = str_to_bool(os.environ.get('CELERY_ALWAYS_EAGER', 'False'))  # False makes tasks run asynchronously
+CELERY_EAGER_PROPAGATES_EXCEPTIONS = str_to_bool(os.environ.get('CELERY_EAGER_PROPAGATES_EXCEPTIONS', 'False'))
+CELERY_RESULT_BACKEND = 'djcelery.backends.database:DatabaseBackend'
+CELERY_IGNORE_RESULT = False
+
+# Slack Settings
+SLACK_BACKEND = os.environ.get('SLACK_BACKEND', 'django_slack.backends.RequestsBackend')
+SLACK_TOKEN = os.environ.get('SLACK_TOKEN', '')
+SLACK_CHANNEL = os.environ.get('SLACK_CHANNEL', '')
+SLACK_ICON_EMOJI = os.environ.get('SLACK_ICON_EMOJI', '')
+SLACK_USERNAME = os.environ.get('SLACK_USERNAME', '')
+
+# Haystack Settings
+HAYSTACK_SEARCH = True
+# Update facet counts from Haystack
+HAYSTACK_FACET_COUNTS = False
+HAYSTACK_URL = "%s://%s:%d" % (os.environ['PRIVATE_PROTOCOL'], os.environ['ELASTIC_HOST'], int(os.environ['ELASTIC_PORT']))
+HAYSTACK_CONNECTIONS = {
+    'default': {
+        'ENGINE': 'mapstory.search.elasticsearch_backend.MapStoryElasticsearchSearchEngine',
+        'URL': HAYSTACK_URL,
+        'INDEX_NAME': 'geonode',
+        'EXCLUDED_INDEXES': ['geonode.layers.search_indexes.LayerIndex'],
+    },
+}
+SKIP_PERMS_FILTER = True
+HAYSTACK_SIGNAL_PROCESSOR = 'mapstory.search.signals.RealtimeSignalProcessor'
+
+# Social Authentication Settings
+ENABLE_SOCIAL_LOGIN = str_to_bool(os.environ['ENABLE_SOCIAL_LOGIN'])
+if ENABLE_SOCIAL_LOGIN:
+    SOCIAL_AUTH_NEW_USER_REDIRECT_URL = '/'
+
+    INSTALLED_APPS += (
+        'social.apps.django_app.default',
+        'provider',
+        'provider.oauth2',
+    )
+
+    AUTHENTICATION_BACKENDS += (
+        'social.backends.google.GoogleOAuth2',
+        'social.backends.facebook.FacebookOAuth2',
+    )
+
+DEFAULT_AUTH_PIPELINE = (
+    'social.pipeline.social_auth.social_details',
+    'social.pipeline.social_auth.social_uid',
+    'social.pipeline.social_auth.auth_allowed',
+    'social.pipeline.social_auth.social_user',
+    'social.pipeline.user.get_username',
+    'social.pipeline.mail.mail_validation',
+    'social.pipeline.social_auth.associate_by_email',
+    'social.pipeline.user.create_user',
+    'social.pipeline.social_auth.associate_user',
+    'social.pipeline.social_auth.load_extra_data',
+    'social.pipeline.user.user_details'
+)
+
+SOCIAL_AUTH_FACEBOOK_KEY = os.environ.get('FACEBOOK_APP_ID','')
+SOCIAL_AUTH_FACEBOOK_SECRET = os.environ.get('FACEBOOK_APP_SECRET','')
+SOCIAL_AUTH_FACEBOOK_SCOPE = ['email']
+SOCIAL_AUTH_FACEBOOK_PROFILE_EXTRA_PARAMS = {
+    'fields': 'id,name,email',
+}
+
+SOCIAL_AUTH_GOOGLE_OAUTH2_KEY = os.environ.get('GOOGLE_OATH2_CLIENT_ID','')
+SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET = os.environ.get('GOOGLE_OATH2_CLIENT_SECRET','')
+
+SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')
+
